LayoutTests/imported/w3c/web-platform-tests/css/css-align/gaps/legacy-gap-aliases-001-expected.txt


PASS grid-row-gap is a legacy name alias for row-gap
PASS grid-column-gap is a legacy name alias for column-gap
PASS grid-gap is a legacy name alias for gap
